pyproject.toml
[project]
name = "public-school-duckdb"
version = "0.1.0"
description = "Public school data dashboard — 학교알리미 Open API → DuckDB pipeline"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
load-data = "schooldata.cli:main"
